Removed settings parameters, combined under one
Compiled the global settings available to the uploader under one
parameter, which is expected to be a JS object, the properties of which
should be the options.

While this change will break backwards compatability for now, it should
be easier to add extra settings in the future while maintaining future
compatability (i.e. being able to upgrade the core uploader package
without modifying the front-end)