Added checkbox required functionality
Moved required checks into a saparate function, isRequired(), which will
now check - if data-required is not equal to yes or no - if it can be
used as a querySelector(). If it can, it will used the checked attribute
as the required status.

Added as a way to incorporate the optional nature of some sections in
certain forms where an upload may be required if the section is
enabled.